mongo compile: implemented where
