+++ conformance/tests/ot_where.py
@@ -16,194 +16,225 @@ def setupGraph(O):
     O.addEdge("vertex2", "vertex4", "friend", gid="edge3")
 
 
-def test_where_eq(O):
-    errors = []
-    setupGraph(O)
+# def test_where_eq(O):
+#     errors = []
+#     setupGraph(O)
 
-    count = 0
-    for i in O.query().V().where(aql.eq("_gid", "vertex3")):
-        count += 1
-        if i['gid'] != "vertex3":
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 1:
-        errors.append(
-            "Fail: O.query().V().where(aql.eq(\"_gid\", \"vertex3\")) %s != %s" %
-            (count, 1))
+#     count = 0
+#     for i in O.query().V().where(aql.eq("_gid", "vertex3")):
+#         count += 1
+#         if i['gid'] != "vertex3":
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 1:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.eq(\"_gid\", \"vertex3\")) %s != %s" %
+#             (count, 1))
 
-    count = 0
-    for i in O.query().V().where(aql.eq("_label", "person")):
-        count += 1
-        if i['label'] != "person":
-            errors.append("Wrong vertex label %s" % (i['label']))
-    if count != 4:
-        errors.append(
-            "Fail: O.query().V().where(aql.eq(\"_label\", \"person\")) %s != %s" %
-            (count, 4))
+#     count = 0
+#     for i in O.query().V().where(aql.eq("_label", "person")):
+#         count += 1
+#         if i['label'] != "person":
+#             errors.append("Wrong vertex label %s" % (i['label']))
+#     if count != 4:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.eq(\"_label\", \"person\")) %s != %s" %
+#             (count, 4))
 
-    count = 0
-    for i in O.query().V().where(aql.eq("occupation", "jedi")):
-        count += 1
-        if i['gid'] not in ["vertex2", "vertex5"]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 2:
-        errors.append(
-            "Fail: O.query().V().where(aql.eq(\"occupation\", \"jedi\")) %s != %s" %
-            (count, 2))
+#     count = 0
+#     for i in O.query().V().where(aql.eq("occupation", "jedi")):
+#         count += 1
+#         if i['gid'] not in ["vertex2", "vertex5"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 2:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.eq(\"occupation\", \"jedi\")) %s != %s" %
+#             (count, 2))
 
-    return errors
-
-
-def test_where_neq(O):
-    errors = []
-    setupGraph(O)
-
-    count = 0
-    for i in O.query().V().where(aql.not_(aql.eq("_gid", "vertex3"))):
-        count += 1
-        if i['gid'] == "vertex3":
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 5:
-        errors.append(
-            "Fail: O.query().V().where(aql.not_(aql.eq(\"_gid\", \"vertex3\"))) %s != %s" %
-            (count, 5))
-
-    count = 0
-    for i in O.query().V().where(aql.not_(aql.eq("_label", "person"))):
-        count += 1
-        if i['label'] == "person":
-            errors.append("Wrong vertex label %s" % (i['label']))
-    if count != 2:
-        errors.append(
-            "Fail: O.query().V().where(aql.not_(aql.eq(\"_label\", \"person\"))) %s != %s" %
-            (count, 2))
-
-    count = 0
-    for i in O.query().V().where(aql.not_(aql.eq("occupation", "jedi"))):
-        count += 1
-        if i['gid'] in ["vertex2", "vertex5"]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 4:
-        errors.append(
-            "Fail: O.query().V().where(aql.not_(aql.eq(\"occupation\", \"jedi\"))) %s != %s" %
-            (count, 4))
-
-    return errors
+#     return errors
 
 
-def test_where_in(O):
-    errors = []
-    setupGraph(O)
+# def test_where_neq(O):
+#     errors = []
+#     setupGraph(O)
 
-    count = 0
-    for i in O.query().V().where((aql.in_("occupation", ["jedi", "sith"]))):
-        count += 1
-        if i['gid'] not in  ["vertex2", "vertex5", "vertex6"]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 3:
-        errors.append(
-            "Fail: O.query().V().where((aql.in_(\"occupation\", [\"jedi\", \"sith\"]))) %s != %s" %
-            (count, 3))
+#     count = 0
+#     for i in O.query().V().where(aql.neq("_gid", "vertex3")):
+#         count += 1
+#         if i['gid'] == "vertex3":
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 5:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.not_(aql.eq(\"_gid\", \"vertex3\"))) %s != %s" %
+#             (count, 5))
 
-    count = 0
-    for i in O.query().V().where((aql.in_("occupation", {"jedi": 0, "sith": 0}))):
-        count += 1
-        if i['gid'] not in  ["vertex2", "vertex5", "vertex6"]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 3:
-        errors.append(
-            "Fail: O.query().V().where((aql.in_(\"occupation\", [\"jedi\", \"sith\"]))) %s != %s" %
-            (count, 3))
+#     count = 0
+#     for i in O.query().V().where(aql.neq("_label", "person")):
+#         count += 1
+#         if i['label'] == "person":
+#             errors.append("Wrong vertex label %s" % (i['label']))
+#     if count != 2:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.not_(aql.eq(\"_label\", \"person\"))) %s != %s" %
+#             (count, 2))
 
-    count = 0
-    for i in O.query().V().where((aql.in_("occupation", 0))):
-        count += 1
-    if count != 0:
-        errors.append(
-            "Fail: O.query().V().where((aql.in_(\"occupation\", [\"jedi\", \"sith\"]))) %s != %s" %
-            (count, 0))
+#     count = 0
+#     for i in O.query().V().where(aql.neq("occupation", "jedi")):
+#         count += 1
+#         if i['gid'] in ["vertex2", "vertex5"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 4:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.not_(aql.eq(\"occupation\", \"jedi\"))) %s != %s" %
+#             (count, 4))
 
-    return errors
-
-
-def test_where_contains(O):
-    errors = []
-    setupGraph(O)
-
-    count = 0
-    for i in O.query().V().where((aql.contains("starships", "x-wing"))):
-        count += 1
-        if i['gid'] not in ["vertex2"]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 1:
-        errors.append(
-            "Fail: O.query().V().where((aql.contains(\"starships\", \"x-wing\"))) %s != %s" %
-            (count, 1))
-
-    count = 0
-    for i in O.query().V().where((aql.contains("_data", "starships"))):
-        count += 1
-        if i['gid'] not in ["vertex1", "vertex2", "vertex6"]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 3:
-        errors.append(
-            "Fail: O.query().V().where((aql.contains(\"starships\", \"x-wing\"))) %s != %s" %
-            (count, 3))
-
-    return errors
+#     return errors
 
 
-def test_where_gt(O):
-    errors = []
-    setupGraph(O)
+# def test_where_in(O):
+#     errors = []
+#     setupGraph(O)
 
-    count = 0
-    for i in O.query().V().where(aql.gt("age", 35)):
-        count += 1
-        if i['gid'] not in ["vertex5", "vertex6"]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 2:
-        errors.append(
-            "Fail: O.query().V().where(aql.gt(\"age\", 35)) %s != %s" %
-            (count, 2))
+#     count = 0
+#     for i in O.query().V().where(aql.in_("occupation", ["jedi", "sith"])):
+#         count += 1
+#         if i['gid'] not in ["vertex2", "vertex5", "vertex6"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 3:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.in_(\"occupation\", [\"jedi\", \"sith\"])) %s != %s" %
+#             (count, 3))
 
-    count = 0
-    for i in O.query().V().where(aql.gte("age", 35)):
-        count += 1
-        if i['gid'] not in ["vertex1", "vertex5", "vertex6"]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 3:
-        errors.append(
-            "Fail: O.query().V().where(aql.gte(\"age\", 35)) %s != %s" %
-            (count, 3))
+#     count = 0
+#     for i in O.query().V().where(aql.in_("occupation", 0)):
+#         count += 1
+#     if count != 0:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.in_(\"occupation\", 0)) %s != %s" %
+#             (count, 0))
 
-    return errors
+#     return errors
 
 
-def test_where_lt(O):
-    errors = []
-    setupGraph(O)
+# def test_where_contains(O):
+#     errors = []
+#     setupGraph(O)
 
-    count = 0
-    for i in O.query().V().where(aql.lt("age", 35)):
-        count += 1
-        if i['gid'] not in ["vertex2"]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 1:
-        errors.append(
-            "Fail: O.query().V().where(aql.gt(\"age\", 35)) %s != %s" %
-            (count, 1))
+#     count = 0
+#     for i in O.query().V().where(aql.contains("starships", "x-wing")):
+#         count += 1
+#         if i['gid'] not in ["vertex2"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 1:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.contains(\"starships\", \"x-wing\")) %s != %s" %
+#             (count, 1))
 
-    count = 0
-    for i in O.query().V().where(aql.lte("age", 35)):
-        count += 1
-        if i['gid'] not in ["vertex1", "vertex2",]:
-            errors.append("Wrong vertex returned %s" % (i))
-    if count != 2:
-        errors.append(
-            "Fail: O.query().V().where(aql.gte(\"age\", 35)) %s != %s" %
-            (count, 2))
+#     return errors
 
-    return errors
+
+# def test_where_gt(O):
+#     errors = []
+#     setupGraph(O)
+
+#     count = 0
+#     for i in O.query().V().where(aql.gt("age", 35)):
+#         count += 1
+#         if i['gid'] not in ["vertex5", "vertex6"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 2:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.gt(\"age\", 35)) %s != %s" %
+#             (count, 2))
+
+#     count = 0
+#     for i in O.query().V().where(aql.gte("age", 35)):
+#         count += 1
+#         if i['gid'] not in ["vertex1", "vertex5", "vertex6"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 3:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.gte(\"age\", 35)) %s != %s" %
+#             (count, 3))
+
+#     return errors
+
+
+# def test_where_lt(O):
+#     errors = []
+#     setupGraph(O)
+
+#     count = 0
+#     for i in O.query().V().where(aql.lt("age", 35)):
+#         count += 1
+#         if i['gid'] not in ["vertex2"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 1:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.lt(\"age\", 35)) %s != %s" %
+#             (count, 1))
+
+#     count = 0
+#     for i in O.query().V().where(aql.lte("age", 35)):
+#         count += 1
+#         if i['gid'] not in ["vertex1", "vertex2"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 2:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.lte(\"age\", 35)) %s != %s" %
+#             (count, 2))
+
+#     return errors
+
+
+# def test_where_and(O):
+#     errors = []
+#     setupGraph(O)
+
+#     count = 0
+#     for i in O.query().V().where(aql.and_(aql.eq("_label", "person"), aql.eq("occupation", "jedi"))):
+#         count += 1
+#         if i['gid'] not in ["vertex2", "vertex5"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 2:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.and_(aql.eq(\"_label\", \"person\"), aql.eq(\"occupation\", \"jedi\"))) %s != %s" %
+#             (count, 2))
+
+#     return errors
+
+
+# def test_where_or(O):
+#     errors = []
+#     setupGraph(O)
+
+#     count = 0
+#     for i in O.query().V().where(aql.or_(aql.eq("occupation", "sith"), aql.eq("occupation", "jedi"))):
+#         count += 1
+#         if i['gid'] not in ["vertex2", "vertex5", "vertex6"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 3:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.or_(aql.eq(\"occupation\", \"sith\"), aql.eq(\"occupation\", \"jedi\"))) %s != %s" %
+#             (count, 3))
+
+#     return errors
+
+
+# def test_where_not(O):
+#     errors = []
+#     setupGraph(O)
+
+#     count = 0
+#     for i in O.query().V().where(aql.not_(aql.eq("_label", "person"))):
+#         count += 1
+#         if i['gid'] not in ["vertex3", "vertex4"]:
+#             errors.append("Wrong vertex returned %s" % (i))
+#     if count != 2:
+#         errors.append(
+#             "Fail: O.query().V().where(aql.not_(aql.eq(\"_label\", \"person\"))) %s != %s" %
+#             (count, 2))
+
+#     return errors
 
 
 def test_where_complex(O):
@@ -221,7 +252,23 @@ def test_where_complex(O):
             errors.append("Wrong vertex returned %s" % (i))
     if count != 1:
         errors.append(
-            "Fail: O.query().V().where(aql.and_(aql.eq(\"_label\", \"person\"), aql.not_(aql.eq(\"occupation\", \"jedi\")))) %s != %s" %
+            "Fail: O.query().V().where(aql.and_(aql.eq(\"_label\", \"person\"), aql.not_(aql.or_(aql.eq(\"occupation\", \"jedi\"), aql.eq(\"occupation\", \"sith\"))))) %s != %s" %
+            (count, 1)
+        )
+
+    count = 0
+    for i in O.query().V().where(aql.not_(
+            aql.and_(
+                aql.eq("_label", "robot"),
+                aql.or_(aql.eq("occupation", "jedi"),
+                        aql.contains("starships", "millennium falcon")))
+            )):
+        count += 1
+        if i['gid'] != "vertex6":
+            errors.append("Wrong vertex returned %s" % (i))
+    if count != 1:
+        errors.append(
+            "Fail: O.query().V().where(aql.not_(aql.and_(aql.eq(\"_label\", \"robot\"), aql.or_(aql.eq(\"occupation\", \"jedi\"),  aql.contains(\"starships\", \"millennium falcon\"))))) %s != %s" %
             (count, 1)
         )
