+++ mongo/compile.go
@@ -1,9 +1,7 @@
 package mongo
 
 import (
-	"context"
 	"fmt"
-	"log"
 	"strings"
 
 	"github.com/bmeg/arachne/aql"
@@ -14,6 +12,28 @@ import (
 	"gopkg.in/mgo.v2/bson"
 )
 
+// Pipeline a set of runnable query operations
+type Pipeline struct {
+	procs     []gdbi.Processor
+	dataType  gdbi.DataType
+	markTypes map[string]gdbi.DataType
+}
+
+// DataType return the datatype
+func (pipe *Pipeline) DataType() gdbi.DataType {
+	return pipe.dataType
+}
+
+// MarkTypes get the mark types
+func (pipe *Pipeline) MarkTypes() map[string]gdbi.DataType {
+	return pipe.markTypes
+}
+
+// Processors gets the list of processors
+func (pipe *Pipeline) Processors() []gdbi.Processor {
+	return pipe.procs
+}
+
 // Compiler is a mongo specific compiler that works with default graph interface
 type Compiler struct {
 	db *Graph
@@ -22,13 +42,6 @@ type Compiler struct {
 // NewCompiler creates a new compiler that runs using the provided GraphInterface
 func NewCompiler(db *Graph) gdbi.Compiler {
 	return &Compiler{db: db}
-}
-
-// Pipeline a set of runnable query operations
-type Pipeline struct {
-	procs     []gdbi.Processor
-	dataType  gdbi.DataType
-	markTypes map[string]gdbi.DataType
 }
 
 // Compile compiles a set of graph traversal statements into a mongo aggregation pipeline
@@ -356,7 +369,9 @@ func (comp *Compiler) Compile(stmts []*aql.GraphStatement) (gdbi.Pipeline, error
 			if lastType != gdbi.VertexData && lastType != gdbi.EdgeData {
 				return &Pipeline{}, fmt.Errorf(`"where" statement is only valid for edge or vertex types not: %s`, lastType.String())
 			}
-			// TODO
+			whereExpr := convertWhereExpression(stmt.Where, false)
+			matchStmt := bson.M{"$match": whereExpr}
+			query = append(query, matchStmt)
 
 		case *aql.GraphStatement_Limit:
 			query = append(query,
@@ -548,156 +563,4 @@ func (comp *Compiler) Compile(stmts []*aql.GraphStatement) (gdbi.Pipeline, error
 
 	procs = append([]gdbi.Processor{&Processor{comp.db, startCollection, query, lastType, markTypes}}, procs...)
 	return &Pipeline{procs, lastType, markTypes}, nil
-}
-
-// DataType return the datatype
-func (pipe *Pipeline) DataType() gdbi.DataType {
-	return pipe.dataType
-}
-
-// MarkTypes get the mark types
-func (pipe *Pipeline) MarkTypes() map[string]gdbi.DataType {
-	return pipe.markTypes
-}
-
-// Processors gets the list of processors
-func (pipe *Pipeline) Processors() []gdbi.Processor {
-	return pipe.procs
-}
-
-// Processor stores the information for a mongo aggregation pipeline
-type Processor struct {
-	db              *Graph
-	startCollection string
-	query           []bson.M
-	dataType        gdbi.DataType
-	markTypes       map[string]gdbi.DataType
-}
-
-// Process runs the mongo aggregation pipeline
-func (proc *Processor) Process(ctx context.Context, man gdbi.Manager, in gdbi.InPipe, out gdbi.OutPipe) context.Context {
-	log.Printf("Running Mongo Processor: %+v", proc.query)
-
-	go func() {
-		session := proc.db.ar.pool.Get()
-		defer proc.db.ar.pool.Put(session)
-		defer close(out)
-
-		initCol := session.DB(proc.db.ar.database).C(proc.startCollection)
-		for t := range in {
-			iter := initCol.Pipe(proc.query).Iter()
-			result := map[string]interface{}{}
-			for iter.Next(&result) {
-				log.Printf("Mongo Pipeline result: %+v", result)
-				select {
-				case <-ctx.Done():
-					return
-				default:
-				}
-
-				switch proc.dataType {
-				case gdbi.CountData:
-					eo := &gdbi.Traveler{}
-					if x, ok := result["count"]; ok {
-						eo.Count = uint32(x.(int))
-					}
-					out <- eo
-
-				case gdbi.SelectionData:
-					selections := map[string]*gdbi.DataElement{}
-					if marks, ok := result["marks"]; ok {
-						if marks, ok := marks.(map[string]interface{}); ok {
-							for k, v := range marks {
-								gid, ok := v.(string)
-								if !ok {
-									log.Printf("Failed to process selection data: %+v", v)
-								}
-								de := &gdbi.DataElement{}
-								switch proc.markTypes[k] {
-								case gdbi.VertexData:
-									v := proc.db.GetVertex(gid, true)
-									if v != nil {
-										de = &gdbi.DataElement{
-											ID:    v.Gid,
-											Label: v.Label,
-											Data:  v.GetDataMap(),
-										}
-									}
-								case gdbi.EdgeData:
-									e := proc.db.GetEdge(gid, true)
-									if e != nil {
-										de = &gdbi.DataElement{
-											ID:    e.Gid,
-											Label: e.Label,
-											From:  e.From,
-											To:    e.To,
-											Data:  e.GetDataMap(),
-										}
-									}
-								}
-								selections[k] = de
-							}
-						}
-					}
-					out <- &gdbi.Traveler{Selections: selections}
-
-				case gdbi.AggregationData:
-					aggs := map[string]*aql.AggregationResult{}
-
-					for k, v := range result {
-						out := &aql.AggregationResult{
-							Buckets: []*aql.AggregationResultBucket{},
-						}
-
-						buckets, ok := v.([]interface{})
-						if !ok {
-							log.Printf("Failed to convert Mongo aggregation result: %+v", v)
-							continue
-						}
-						for _, bucket := range buckets {
-							bucket, ok := bucket.(map[string]interface{})
-							if !ok {
-								log.Printf("Failed to convert Mongo aggregation result: %+v", bucket)
-								continue
-							}
-							term := protoutil.WrapValue(bucket["_id"])
-							count, ok := bucket["count"].(int)
-							if !ok {
-								log.Printf("failed to cast count result to integer: %v", bucket)
-								continue
-							}
-							out.Buckets = append(out.Buckets, &aql.AggregationResultBucket{Key: term, Value: float64(count)})
-						}
-						aggs[k] = out
-					}
-					out <- &gdbi.Traveler{Aggregations: aggs}
-
-				default:
-					de := &gdbi.DataElement{}
-					if x, ok := result["_id"]; ok {
-						de.ID = x.(string)
-					}
-					if x, ok := result["label"]; ok {
-						de.Label = x.(string)
-					}
-					if x, ok := result["data"]; ok {
-						de.Data = x.(map[string]interface{})
-					}
-					if x, ok := result["to"]; ok {
-						de.To = x.(string)
-					}
-					if x, ok := result["from"]; ok {
-						de.From = x.(string)
-					}
-					out <- t.AddCurrent(de)
-				}
-			}
-			if err := iter.Err(); err != nil {
-				log.Println("Mongo traversal error:", err)
-				continue
-			}
-		}
-	}()
-
-	return ctx
 }
